hopr_chain_indexer::block

Struct Indexer

pub struct Indexer<T, U, Db>
where
    T: HoprIndexerRpcOperations + Send + 'static,
    U: ChainLogHandler + Send + 'static,
    Db: HoprDbGeneralModelOperations + HoprDbInfoOperations + HoprDbLogOperations + Clone + Send + Sync + 'static,
{ /* private fields */ }

Indexer

Accepts the RPC operational functionality hopr_chain_rpc::HoprIndexerRpcOperations and provides the indexing operation resulting in and output of hopr_chain_types::chain_events::SignificantChainEvent streamed outside the indexer by the unbounded channel.

The roles of the indexer:

prime the RPC endpoint

request an RPC stream of changes to process

process block and log stream

ensure finalization by postponing processing until the head is far enough

store relevant data into the DB

pass the processing on to the business logic

Implementations

impl<T, U, Db> Indexer<T, U, Db>
where T: HoprIndexerRpcOperations + Sync + Send + 'static, U: ChainLogHandler + Send + Sync + 'static, Db: HoprDbGeneralModelOperations + HoprDbInfoOperations + HoprDbLogOperations + Clone + Send + Sync + 'static,

pub fn new( rpc: T, db_processor: U, db: Db, cfg: IndexerConfig, egress: Sender<SignificantChainEvent>, ) -> Self

pub fn without_panic_on_completion(self) -> Self

Disables the panic on completion.

pub async fn start(self) -> Result<AbortHandle>
where T: HoprIndexerRpcOperations + 'static, U: ChainLogHandler + 'static, Db: HoprDbGeneralModelOperations + HoprDbInfoOperations + HoprDbLogOperations + Clone + Send + Sync + 'static,

pub async fn pre_start(&self) -> Result<()>

pub async fn download_snapshot(&self) -> Result<SnapshotInfo>

Downloads and installs a database snapshot for faster initial synchronization.

This method coordinates the snapshot download process by:

Validating the indexer configuration

Creating a snapshot manager instance

Downloading and extracting the snapshot to the data directory

Snapshots allow new nodes to quickly synchronize with the network by downloading pre-built database files instead of fetching all historical logs from scratch.

Returns

Ok(SnapshotInfo) containing details about the downloaded snapshot

Err(CoreEthereumIndexerError::SnapshotError) if validation or download fails

Prerequisites

Configuration must be valid (proper URL format, data directory set)

Sufficient disk space must be available

Network connectivity to the snapshot URL